Create contrib directory and move swann to u_arizona swann analysis path within contrib
