+++ src/reformatters/contrib/u_arizona/swann/analysis/templates/latest.zarr/snow_water_equivalent/zarr.json
