+++ src/reformatters/contrib/u_arizona/swann/analysis/templates/latest.zarr/spatial_ref/zarr.json
